Improved Request Handling
This commit improves upon the requests handled by the SPII device. It abstract a further axis class in order to clean up the requests for dynamic axis handling within the software.

+++ src/library_SPIIMotionController/spii_poll_machine.cpp
@@ -25,7 +25,7 @@ void SPIIPollMachine::addRequestToQueue(const AbstractRequestPtr request, const 
     timeoutMap[request->getTupleDescription()] = dataPair;
 
     std::map<common::TupleECMData,pollingTimeout>::iterator it = timeoutMap.begin();
-    int currentTimeout = it->second.second;
+    double currentTimeout = it->second.second;
     std::advance(it,1);
     for (; it!=timeoutMap.end(); ++it)
     {
